Put Fragments and model classes in respective packages
