+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/MainActivity.java
@@ -24,6 +24,7 @@ public class MainActivity extends AppCompatActivity {
     ActionBarDrawerToggle actionBarDrawerToggle;
 
 
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
