+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/fragments/MealPlannerFragment.java
@@ -1,4 +1,4 @@
-package com.git_er_done.cmput301f22t06_team_project;
+package com.git_er_done.cmput301f22t06_team_project.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.git_er_done.cmput301f22t06_team_project.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ShoppingListFragment#newInstance} factory method to
+ * Use the {@link MealPlannerFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ShoppingListFragment extends Fragment {
+public class MealPlannerFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class ShoppingListFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public ShoppingListFragment() {
+    public MealPlannerFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class ShoppingListFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ShoppingListFragment.
+     * @return A new instance of fragment MealPlannerFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static ShoppingListFragment newInstance(String param1, String param2) {
-        ShoppingListFragment fragment = new ShoppingListFragment();
+    public static MealPlannerFragment newInstance(String param1, String param2) {
+        MealPlannerFragment fragment = new MealPlannerFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class ShoppingListFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_shopping_list, container, false);
+        return inflater.inflate(R.layout.fragment_meal_planner, container, false);
     }
 }
