+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/fragments/ShoppingListFragment.java
@@ -1,4 +1,4 @@
-package com.git_er_done.cmput301f22t06_team_project;
+package com.git_er_done.cmput301f22t06_team_project.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.git_er_done.cmput301f22t06_team_project.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link RecipesFragment#newInstance} factory method to
+ * Use the {@link ShoppingListFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class RecipesFragment extends Fragment {
+public class ShoppingListFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class RecipesFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public RecipesFragment() {
+    public ShoppingListFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class RecipesFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment RecipesFragment.
+     * @return A new instance of fragment ShoppingListFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static RecipesFragment newInstance(String param1, String param2) {
-        RecipesFragment fragment = new RecipesFragment();
+    public static ShoppingListFragment newInstance(String param1, String param2) {
+        ShoppingListFragment fragment = new ShoppingListFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class RecipesFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_recipes, container, false);
+        return inflater.inflate(R.layout.fragment_shopping_list, container, false);
     }
 }
