+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/models/MealPlan.java
@@ -1,4 +1,4 @@
-package com.git_er_done.cmput301f22t06_team_project;
+package com.git_er_done.cmput301f22t06_team_project.models;
 
 import java.util.ArrayList;
