+++ android_studio_project/app/src/main/res/layout/fragment_ingredient.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IngredientsFragment">
+    tools:context=".fragments.IngredientsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
