+++ android_studio_project/app/src/main/res/layout/fragment_shopping_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ShoppingListFragment">
+    tools:context=".fragments.ShoppingListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
