+++ android_studio_project/app/src/main/res/navigation/nav_graph.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/recipesFragment"
-        android:name="com.git_er_done.cmput301f22t06_team_project.RecipesFragment"
+        android:name="com.git_er_done.cmput301f22t06_team_project.fragments.RecipesFragment"
         android:label="fragment_recipes"
         tools:layout="@layout/fragment_recipes" >
     </fragment>
     <fragment
         android:id="@+id/ingredientsFragment"
-        android:name="com.git_er_done.cmput301f22t06_team_project.IngredientsFragment"
+        android:name="com.git_er_done.cmput301f22t06_team_project.fragments.IngredientsFragment"
         android:label="fragment_ingredient"
         tools:layout="@layout/fragment_ingredient" >
     </fragment>
     <fragment
         android:id="@+id/mealPlannerFragment"
-        android:name="com.git_er_done.cmput301f22t06_team_project.MealPlannerFragment"
+        android:name="com.git_er_done.cmput301f22t06_team_project.fragments.MealPlannerFragment"
         android:label="fragment_meal_planner"
         tools:layout="@layout/fragment_meal_planner" >
     </fragment>
     <fragment
         android:id="@+id/shoppingListFragment"
-        android:name="com.git_er_done.cmput301f22t06_team_project.ShoppingListFragment"
+        android:name="com.git_er_done.cmput301f22t06_team_project.fragments.ShoppingListFragment"
         android:label="fragment_shopping_list"
         tools:layout="@layout/fragment_shopping_list" >
